+++ square.cpp
@@ -8,7 +8,7 @@
 Square::Square() {
     std::cout << "\nDefault constructor called SQUARE " << this;
 }
-Square::Square(int px, int py, int pvelx, int pvely, int psizex, int psizey, Color pcolor, bool pmovable) {
+Square::Square(double px, double py, double pvelx, double pvely, double psizex, double psizey, Color pcolor, bool pmovable) {
     std::cout << "\nConstructor called SQUARE " << this;
     pos.x = px, pos.y = py, vel.x = pvelx, vel.y = pvely, size.x = psizex, size.y = psizey, color = pcolor, movable = pmovable;
 }
